requirements.txt
Pillow
tensorflow
keras
numpy
pandas
matplotlib
scikit-learn